src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
anyio==3.6.2
    # via httpcore
beautifulsoup4==4.11.1
    # via -r requirements.in
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
h11==0.14.0
    # via httpcore
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==0.16.3
    # via httpx
httpx[http2]==0.23.3
    # via -r requirements.in
hyperframe==6.0.1
    # via h2
idna==3.4
    # via
    #   anyio
    #   rfc3986
rfc3986[idna2008]==1.5.0
    # via httpx
ruamel-yaml==0.17.21
    # via -r requirements.in
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
soupsieve==2.3.2.post1
    # via beautifulsoup4
